requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
customtkinter==5.0.3
darkdetect==0.8.0
greenlet==2.0.1
idna==3.4
numpy==1.24.0
opencv-python==4.6.0.66
pandas==1.5.2
pip==21.2.3
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
setuptools==57.4.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
urllib3==1.26.13
matplotlib==3.7.0
Pillow==9.3.0
